Hide the implementation files in a special directory; rename files according to the classes
