+++ cmd/download_headers.cpp
@@ -19,12 +19,12 @@
 #include <thread>
 
 #include <CLI/CLI.hpp>
+#include <node/silkworm/downloader/internals/header_retrieval.hpp>
 
 #include <silkworm/common/directories.hpp>
 #include <silkworm/common/log.hpp>
-#include <silkworm/downloader/HeaderLogic.hpp>
-#include <silkworm/downloader/SentryClient.hpp>
-#include <silkworm/downloader/stage1.hpp>
+#include <silkworm/downloader/block_provider.hpp>
+#include <silkworm/downloader/sentry_client.hpp>
 
 using namespace silkworm;
 
@@ -41,10 +41,12 @@ int main(int argc, char* argv[]) {
     string temporary_file_path = ".";
     string sentry_addr = "127.0.0.1:9091";
 
-    app.add_option("--chaindata", db_path, "Path to the chain database", true)->check(CLI::ExistingDirectory);
-    app.add_option("--chain", chain_name, "Network name", true)->needs("--chaindata");
+    app.add_option("--chaindata", db_path, "Path to the chain database", true)
+        ->check(CLI::ExistingDirectory);
+    app.add_option("--chain", chain_name, "Network name", true)
+        ->needs("--chaindata");
     app.add_option("-s,--sentryaddr", sentry_addr, "address:port of sentry", true);
-    //  todo ->check?
+        //  todo ->check?
     app.add_option("-f,--filesdir", temporary_file_path, "Path to a temp files dir", true)
         ->check(CLI::ExistingDirectory);
