+++ node/silkworm/downloader/block_provider.hpp
@@ -13,8 +13,8 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-#ifndef SILKWORM_STAGE1_HPP
-#define SILKWORM_STAGE1_HPP
+#ifndef SILKWORM_BLOCK_PROVIDER_HPP
+#define SILKWORM_BLOCK_PROVIDER_HPP
 
 #include <chrono>
 
@@ -22,11 +22,11 @@
 #include <silkworm/concurrency/active_component.hpp>
 #include <silkworm/concurrency/containers.hpp>
 
-#include "DbTx.hpp"
-#include "SentryClient.hpp"
-#include "Singleton.hpp"
-#include "Types.hpp"
 #include "messages/Message.hpp"
+#include "sentry_client.hpp"
+#include "silkworm/downloader/internals/DbTx.hpp"
+#include "silkworm/downloader/internals/singleton.hpp"
+#include "silkworm/downloader/internals/types.hpp"
 
 namespace silkworm {
 
@@ -58,4 +58,4 @@ class BlockProvider : public ActiveComponent {  // but also an active component 
 
 }  // namespace silkworm
 
-#endif  // SILKWORM_STAGE1_HPP
+#endif  // SILKWORM_BLOCK_PROVIDER_HPP
