+++ node/silkworm/downloader/internals/DbTx.hpp
@@ -23,7 +23,7 @@
 #include <silkworm/db/tables.hpp>
 #include <silkworm/db/util.hpp>
 
-#include "Types.hpp"
+#include "types.hpp"
 
 using namespace silkworm;
