+++ node/silkworm/downloader/internals/singleton.hpp
