+++ node/silkworm/downloader/packets/BlockBodiesPacket.hpp
@@ -17,7 +17,7 @@
 #ifndef SILKWORM_BLOCKBODIESPACKET_HPP
 #define SILKWORM_BLOCKBODIESPACKET_HPP
 
-#include <silkworm/downloader/Types.hpp>
+#include <silkworm/downloader/internals/types.hpp>
 
 namespace silkworm {
