+++ node/silkworm/downloader/rpc/SendMessageById.hpp
@@ -17,7 +17,7 @@
 #ifndef SILKWORM_SENDMESSAGEBYID_HPP
 #define SILKWORM_SENDMESSAGEBYID_HPP
 
-#include <silkworm/downloader/SentryClient.hpp>
+#include <silkworm/downloader/sentry_client.hpp>
 
 namespace silkworm::rpc {
